Update README.md
Clarity in the project aims.  In other words, this project is somewhat like paid for Archviz tools, but open source and designed with users at the core.  The aim would be clean UI, and intuitive UX, with a "white paper" feel to the interface like existing archviz, Design and CAD tools.